// src/types/partner/contact-client.ts
export interface IContactClient {
  id: string
  id_client_location: string
  id_client?: string
  nome_contato: string
  cargo?: string
  email?: string
  phone?: string
  telefone_secundario?: string
  linkedin?: string
  data_nascimento?: string
  status_contato?: boolean
  data_registro_contato?: string
  notas?: string
}